+++ elfi/methods/parameter_inference.py
@@ -1126,6 +1126,27 @@ class BayesianOptimization(ParameterInference):
         """
         return vis.plot_discrepancy(self.target_model, self.parameter_names, axes=axes, **kwargs)
 
+    def plot_gp(self, axes=None, resol=50, const=None, bounds=None, **kwargs):
+        """Plot pairwise relationships as a matrix with parameters vs. discrepancy.
+
+        Parameters
+        ----------
+        axes : matplotlib.axes.Axes, optional
+        resol : int, optional
+            Resolution of the plotted grid.
+        const : np.array, optional
+            Values for parameters in plots where held constant. Defaults to minimum evidence.
+        bounds: list of tuples, optional
+            List of tuples for axis boundaries.
+
+        Returns
+        -------
+        axes : np.array of plt.Axes
+
+        """
+        return vis.plot_gp(self.target_model, self.parameter_names, axes,
+                           resol, const, bounds, **kwargs)
+
 
 class BOLFI(BayesianOptimization):
     """Bayesian Optimization for Likelihood-Free Inference (BOLFI).