+++ elfi/visualization/visualization.py
@@ -397,3 +397,55 @@ def plot_discrepancy(gp, parameter_names, axes=None, **kwargs):
         axes[idx].set_axis_off()
 
     return axes
+
+
+def plot_gp(gp, parameter_names, axes=None, resol=50, const=None, bounds=None, **kwargs):
+    """Plot pairwise relationships as a matrix with parameters vs. discrepancy.
+
+    Parameters
+    ----------
+    gp : GPyRegression, required
+    parameter_names : list, required
+        Parameter names in format ['mu_0', 'mu_1', ..]
+    axes : plt.Axes or arraylike of plt.Axes
+    resol : int, optional
+        Resolution of the plotted grid.
+    const : np.array, optional
+        Values for parameters in plots where held constant. Defaults to minimum evidence.
+    bounds: list of tuples, optional
+        List of tuples for axis boundaries.
+
+    Returns
+    -------
+    axes : np.array of plt.Axes
+
+    """
+    n_plots = gp.input_dim
+    shape = (n_plots, n_plots)
+    axes, kwargs = _create_axes(axes, shape, **kwargs)
+
+    x_evidence = gp.X
+    y_evidence = gp.Y
+    if const is None:
+        const = x_evidence[np.argmin(y_evidence), :]
+    bounds = bounds or gp.bounds
+
+    for ix in range(n_plots):
+        for jy in range(n_plots):
+            if ix == jy:
+                axes[jy, ix].scatter(x_evidence[:, ix], y_evidence)
+                axes[jy, ix].set_xlim(bounds[ix])
+                axes[jy, ix].set_ylabel('Discrepancy')
+            else:
+                x1 = np.linspace(bounds[ix][0], bounds[ix][1], resol)
+                y1 = np.linspace(bounds[jy][0], bounds[jy][1], resol)
+                x, y = np.meshgrid(x1, y1)
+                predictors = np.tile(const, (resol * resol, 1))
+                predictors[:, ix] = x.ravel()
+                predictors[:, jy] = y.ravel()
+                z = gp.predict_mean(predictors).reshape(resol, resol)
+                axes[jy, ix].contourf(x, y, z)
+                axes[jy, ix].set_ylabel(parameter_names[jy])
+            axes[jy, ix].set_xlabel(parameter_names[ix])
+
+    return axes