+++ CHANGELOG.rst
@@ -4,8 +4,8 @@ Changelog
 dev
 ---
 - Fix bug in plot_discrepancy for more than 6 parameters
-- Implement plot_gp for BayesianOptimization classes for plotting parameters vs. contours
-  in case when we have more than 2 parameters
+- Implement plot_gp for BayesianOptimization classes for plotting discrepancies
+  and pair-wise contours in case when we have arbitrary number of parameters (up to 8)
 
 0.7.3 (2018-08-30)
 ------------------