+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ dev
 ---
 - Fix bug in plot_discrepancy for more than 6 parameters
 - Implement plot_gp for BayesianOptimization classes for plotting discrepancies
-  and pair-wise contours in case when we have arbitrary number of parameters (up to 8)
+  and pair-wise contours in case when we have arbitrary number of parameters (from 1 up to 8)
 
 0.7.3 (2018-08-30)
 ------------------