+++ CHANGELOG.rst
@@ -1,8 +1,12 @@
 Changelog
 =========
 
-dev
----
+
+- Add sampler option `algorithm` for bolfi-posterior-sampling
+- Add a check whether the option given for `algorithm` is one if the implemented samplers
+- Add metropolis sampler `algorithm=metropolis` for bolfi-posterior-sampling
+- Add option `warmup` to metropolis-sampler
+- Add a small test of metropolis-sampler
 - Fix bug in plot_discrepancy for more than 6 parameters
 - Implement plot_gp for BayesianOptimization classes for plotting discrepancies
   and pair-wise contours in case when we have arbitrary number of parameters