+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Changelog
 - Add option `warmup` to metropolis-sampler
 - Add a small test of metropolis-sampler
 - Fix bug in plot_discrepancy for more than 6 parameters
+- Implement plot_gp for BayesianOptimization classes for plotting discrepancies
+  and pair-wise contours in case when we have arbitrary number of parameters
 
 0.7.3 (2018-08-30)
 ------------------